+++ src/main/java/fr/estecka/variantscit/VariantsCitMod.java
@@ -36,7 +36,7 @@ implements ClientModInitializer, PreparableModelLoadingPlugin<Map<Item,VariantMa
 
 		ModuleRegistry.Register(Identifier.ofVanilla("axolotl_variant"), new AxolotlBucketModule());
 		ModuleRegistry.Register(Identifier.ofVanilla("custom_data"), CustomVariantModule::new, CustomVariantModule.CODEC);
-		ModuleRegistry.Register(Identifier.ofVanilla("custom_name"), SimpleCustomName::new, SimpleCustomName.CODEC);
+		ModuleRegistry.Register(Identifier.ofVanilla("custom_name"), CustomNameModule::new, CustomNameModule.CODEC);
 		ModuleRegistry.Register(Identifier.ofVanilla("instrument"), new GoatHornModule());
 		ModuleRegistry.Register(Identifier.ofVanilla("jukebox_playable"), new MusicDiscModule());
 		ModuleRegistry.Register(Identifier.ofVanilla("potion_effect"), new PotionEffectModule());
